INTEGRATION: Integration of 5327,5329,5331,5336,5340,5343,5351,5354 from child to parent (branch spec: dev_to_dt_dev).
[git-p4: depot-paths = "//open/dev/": change = 5355]

+++ farrago/src/net/sf/farrago/ddl/DdlValidator.java
@@ -756,10 +756,13 @@ public class DdlValidator extends FarragoCompoundAllocation
         checkValidationExcnQueue();
 
         if (ddlStmt instanceof DdlDropStmt) {
+            checkInUse(ddlStmt.getModelElement().refMofId());
             // Process deletions until a fixpoint is reached, using MDR events
             // to implement RESTRICT/CASCADE.
             while (!deleteQueue.isEmpty()) {
                 RefObject refObj = (RefObject) deleteQueue.iterator().next();
+                checkInUse(refObj.refMofId());
+
                 deleteQueue.remove(refObj);
                 if (!schedulingMap.containsKey(refObj.refMofId())) {
                     if (tracer.isLoggable(Level.FINE)) {
@@ -784,6 +787,14 @@ public class DdlValidator extends FarragoCompoundAllocation
             replacementTarget = findDuplicate(ddlStmt.getModelElement());
 
             if (replacementTarget != null) {
+                if (stmtValidator.getDdlLockManager().isObjectInUse(replacementTarget.refMofId())) {
+                    throw FarragoResource.instance().ValidatorReplacedObjectInUse.ex(
+                            getRepos().getLocalizedObjectName(
+                                null,
+                                replacementTarget.getName(),
+                                replacementTarget.refClass()));
+                }
+
                 // if this is CREATE OR REPLACE and we've encountered an object to be
                 // replaced, save its dependencies for revalidation and delete it.
                 deleteReplacementTarget(ddlStmt.getModelElement());
@@ -1419,6 +1430,29 @@ public class DdlValidator extends FarragoCompoundAllocation
        // intentionally empty
     }
 
+    /**
+     * Checks if an object is in use by a running statement.  If so,
+     * throw a deferred exception.  Used only in the DROP case.
+     * CREATE OR REPLACE throws a different exception.
+     * @param mofId Object MOFID being dropped
+     */
+    private void checkInUse(final String mofId) {
+        if (stmtValidator.getDdlLockManager().isObjectInUse(mofId)) {
+            enqueueValidationExcn(
+                new DeferredException() {
+                    EigenbaseException getException()
+                    {
+                        CwmModelElement droppedElement =
+                            (CwmModelElement) getRepos().getMdrRepos().getByMofId(mofId);
+                        throw FarragoResource.instance().ValidatorDropObjectInUse.ex(
+                                getRepos().getLocalizedObjectName(
+                                    droppedElement,
+                                    droppedElement.refClass()));
+                    }
+                });
+        }
+    }
+    
     //~ Inner Classes ---------------------------------------------------------
 
     /**
